Moved node events to a shared events packages to prevent circular dependencies.
